+++ src/resources/parsers/body1.ts
@@ -1,4 +1,3 @@
-import * as THREE from 'three';
 import { WORLD_SCALE, PolygonType } from '../../utils/lba';
 import { computeBoundingBox } from './body2';
 
@@ -25,12 +24,11 @@ export const parseModelLBA1 = (resource, index, bodyProps) => {
 
     let offset = loadVertices(obj, data, 0x1A);
     offset = loadBones(obj, data, offset);
-    offset = loadShades(obj, data, offset);
+    offset = loadNormals(obj, data, offset);
     offset = loadPolygons(obj, data, offset);
     offset = loadLines(obj, data, offset);
     loadSpheres(obj, data, offset);
 
-    normalizeNormals(obj);
     computeBoundingBox(obj, bodyProps);
 
     return obj;
@@ -47,7 +45,6 @@ function loadVertices(object: any, data: DataView, offset: number) {
             z: data.getInt16(offset + 4, true) * WORLD_SCALE,
             bone: 0
         });
-        object.normals.push(new THREE.Vector3());
         offset += 6;
     }
     return offset;
@@ -64,12 +61,16 @@ function loadBones(object: any, data: DataView, offset: number) {
         const baseElement = data.getInt16(offset + 6, true);
         offset += 8;
 
-        // skip data
+        const normalsSize = data.getInt16(offset + 10, true);
+        // skip default body pose
         offset += 30;
 
         const bone = {
             parent: (baseElement === -1) ? 0xffff : baseElement / 38,
             vertex: basePoint,
+            firstVertex: firstPoint,
+            vertexSize: pointsSize,
+            normalsSize,
         };
 
         for (let j = 0; j < pointsSize; j += 1) {
@@ -81,16 +82,16 @@ function loadBones(object: any, data: DataView, offset: number) {
     return offset;
 }
 
-function loadShades(object: any, data: DataView, offset: number) {
-    object.shadesSize = data.getUint16(offset, true);
+function loadNormals(object: any, data: DataView, offset: number) {
+    object.normalsSize = data.getUint16(offset, true);
     offset += 2;
-    object.shades = [];
-    for (let i = 0; i < object.shadesSize; i += 1) {
-        object.shades.push({
-            unk1: data.getInt16(offset, true),
-            unk2: data.getInt16(offset + 2, true),
-            unk3: data.getInt16(offset + 4, true),
-            unk4: data.getInt16(offset + 6, true),
+    object.normals = [];
+    for (let i = 0; i < object.normalsSize; i += 1) {
+        object.normals.push({
+            x: data.getInt16(offset, true),
+            y: data.getInt16(offset + 2, true),
+            z: data.getInt16(offset + 4, true),
+            colour:  Math.floor((data.getInt16(offset + 6, true) & 0x00FF) / 16)
         });
         offset += 8;
     }
@@ -118,17 +119,17 @@ function loadPolygons(object: any, data: DataView, offset: number) {
     for (let i = 0; i < object.polygonsSize; i += 1) {
         const renderType = data.getUint8(offset);
         const numVertex = data.getUint8(offset + 1);
-        const colour = data.getUint8(offset + 2);
         const hasExtra = !!((renderType & 0x40));
         const hasTransparency = (renderType === 2);
-        offset += 4;
+        offset += 2;
 
         const poly = {
             renderType,
             polyType: PolyTypeMapping[renderType & 0x0F],
             vertex: [],
-            colour: Math.floor(colour / 16),
-            intensity: colour % 16,
+            normals: [],
+            colour: 0,
+            intensity: 0,
             u: [],
             v: [],
             tex: 0,
@@ -138,71 +139,33 @@ function loadPolygons(object: any, data: DataView, offset: number) {
             hasTransparency
         };
 
-        if (renderType >= 7 && !(renderType >= 9)) {
-            // poly.intensity = data.getInt16(offset, true);
+        const intensity = data.getUint8(offset);
+        poly.colour = Math.floor(intensity / 16);
+        poly.intensity = intensity % 16;
+        offset += 2;
+
+        let normal: number = null;
+        if (renderType === 7 || renderType === 8) { // FLAT, GRANIT
+            normal = data.getInt16(offset, true);
             offset += 2;
         }
 
         // vertex block
         for (let k = 0; k < numVertex; k += 1) {
-            if (renderType >= 9) {
-                // poly.intensity = data.getInt16(offset, true);
+            if (renderType >= 9) { // GOURAUD
+                normal = data.getInt16(offset, true);
                 offset += 2;
             }
-            poly.vertex[k] = data.getUint16(offset, true) / 6;
+            const vertexIndex = data.getUint16(offset, true) / 6;
             offset += 2;
-        }
-        const normal = getFaceNormal(object, poly);
-        for (const vertex of poly.vertex) {
-            object.normals[vertex].add(normal);
+
+            poly.vertex.push(vertexIndex);
+            poly.normals.push(normal);
         }
 
         object.polygons.push(poly);
     }
     return offset;
-}
-
-const U = new THREE.Vector3();
-const V = new THREE.Vector3();
-const P1 = new THREE.Vector3();
-const N = new THREE.Vector3();
-
-function getFaceNormal(object, poly) {
-    const pt0 = poly.vertex[0];
-    const pt1 = poly.vertex[1];
-    const pt2 = poly.vertex[2];
-    const vert0 = object.vertices[pt0];
-    const vert1 = object.vertices[pt1];
-    const vert2 = object.vertices[pt2];
-    P1.copy(vert0).add(getBoneOffset(object, vert0.bone));
-    U.copy(vert1).add(getBoneOffset(object, vert1.bone)).sub(P1);
-    V.copy(vert2).add(getBoneOffset(object, vert2.bone)).sub(P1);
-    N.set(
-        (U.y * V.z) - (U.z * V.y),
-        (U.z * V.x) - (U.x * V.z),
-        (U.x * V.y) - (U.y * V.x),
-    );
-    return N.normalize();
-}
-
-const B = new THREE.Vector3();
-
-function getBoneOffset(object, bone) {
-    B.set(0, 0, 0);
-
-    while (bone !== 0xFFFF) {
-        const vert = object.vertices[bone];
-        B.add(vert);
-        bone = object.bones[bone].parent;
-    }
-
-    return B;
-}
-
-function normalizeNormals(object) {
-    for (const normal of object.normals) {
-        normal.normalize();
-    }
 }
 
 function loadLines(object: any, data: DataView, offset: number) {